move getAdUrlByRegion to libs
